Arquitectura de proyecto PF

+++ PLA.Api/ApplicationBusinessRules/DTOs/UsuarioDTO.cs
@@ -7,7 +7,7 @@ public class UsuarioDTO
     public string ApellidoPaterno { get; set; }
     public string ApellidoMaterno { get; set; }
     public DateTime FechaNacimiento { get; set; }
-    public byte Telefono { get; set; }
+    public int Telefono { get; set; }
     public string Password { get; set; }
     public bool Estatus { get; set; }
 }
