+++ PLA.Api/EnterpriseBusinessRules/Entities/POCOs/Usuario.cs
@@ -6,7 +6,7 @@ public class Usuario
     public string ApellidoPaterno { get; set; }
     public string ApellidoMaterno { get; set; }
     public DateTime FechaNacimiento { get; set; }
-    public byte Telefono { get; set; }
+    public int Telefono { get; set; }
     public string Password { get; set; }
     public bool Estatus { get; set; }
     public virtual ICollection<Pedido> Pedidos { get; set; }
